debugging for 3 days

+++ app/utils/function.js
@@ -1,7 +1,7 @@
 const path = require("path");
 const createError = require("http-errors");
 const jwt = require("jsonwebtoken");
-const fs = require("fs")
+const fs = require("fs");
 const { userModel } = require("../model/users");
 const {
   ACCESS_TOKEN_SECRET_KEY,
@@ -71,9 +71,11 @@ async function verfiyRefreshToken(token) {
     });
   });
 }
-function deleteFileInPublic(fileAddress){
-  const pathFile = path.join(__dirname,"..","..","public",fileAddress);
-  fs.unlinkSync(pathFile)
+function deleteFileInPublic(fileAddress) {
+  if (fileAddress) {
+    const pathFile = path.join(__dirname, "..", "..", "public", fileAddress);
+    fs.unlinkSync(pathFile);
+  }
 }
 
 module.exports = {
@@ -81,5 +83,5 @@ module.exports = {
   signAccessToken,
   signRefreshToken,
   verfiyRefreshToken,
-  deleteFileInPublic
+  deleteFileInPublic,
 };
